requirements.txt
fastapi
uvicorn
openai
requests
pyautogui
python-dotenv
pillow
